beman-tidy: add skel for all Beman Standard checks
