+++ tools/beman-tidy/lib/checks/beman_standard/changelog.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO CHANGELOG.TITLE
+# TODO CHANGELOG.FORMAT
