+++ tools/beman-tidy/lib/checks/beman_standard/cmake.py
@@ -1,0 +1,17 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO CMAKE.DEFAULT
+# TODO CMAKE.USE_FETCH_CONTENT
+# TODO CMAKE.PROJECT_NAME
+# TODO CMAKE.PASSIVE_PROJECTS
+# TODO CMAKE.LIBRARY_NAME
+# TODO CMAKE.LIBRARY_ALIAS
+# TODO CMAKE.TARGET_NAMES
+# TODO CMAKE.PASSIVE_TARGETS
+# TODO CMAKE.SKIP_TESTS
+# TODO CMAKE.SKIP_EXAMPLES
+# TODO CMAKE.AVOID_PASSTHROUGHS
