+++ tools/beman-tidy/lib/checks/beman_standard/directory.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO DIRECTORY.INTERFACE_HEADERS
+# TODO DIRECTORY.IMPLEMENTATION_HEADERS
+# TODO DIRECTORY.SOURCES
+# TODO DIRECTORY.TESTS
+# TODO DIRECTORY.EXAMPLES
+# TODO DIRECTORY.DOCS
+# TODO DIRECTORY.PAPERS
