+++ tools/beman-tidy/lib/checks/beman_standard/file.py
@@ -1,0 +1,10 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO FILE.NAMES
+# TODO FILE.TEST_NAMES
+# TODO FILE.LICENSE_ID
+# TODO FILE.COPYRIGHT
