+++ tools/beman-tidy/lib/checks/beman_standard/general.py
@@ -1,0 +1,8 @@
+
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+# TODO LIBRARY.NAMES
+# TODO REPOSITORY.NAME
+# TODO REPOSITORY.CODEOWNERS
+# TODO REPOSITORY.DISALLOW_GIT_SUBMODULES
