+++ tools/beman-tidy/lib/checks/beman_standard/licence.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO LICENSE.APPROVED
+# TODO LICENSE.APACHE_LLVM
+# TODO LICENSE.CRITERIA
