+++ tools/beman-tidy/lib/checks/beman_standard/readme.py
@@ -1,0 +1,10 @@
+#!/usr/bin/python3
+# SPDX-License-Identifier: 2.0 license with LLVM exceptions
+
+from ..abstract.generic_check import BSCheck
+
+
+# TODO README.TITLE
+# TODO README.PURPOSE
+# TODO README.IMPLEMENTS
+# TODO README.LIBRARY_STATUS
